Minor change.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@3934 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/i_keithley2600.c
@@ -142,9 +142,9 @@ mxi_keithley2600_open( MX_RECORD *record )
 
 	snprintf( format, sizeof(format),
 		"Keithley Instruments Inc., Model %%%ds %%%ds %%%ds",
-			sizeof(keithley2600->model_name),
-			sizeof(keithley2600->serial_number),
-			sizeof(keithley2600->firmware_version) );
+			(int) sizeof(keithley2600->model_name),
+			(int) sizeof(keithley2600->serial_number),
+			(int) sizeof(keithley2600->firmware_version) );
 
 	num_items = sscanf( response, format,
 			keithley2600->model_name,
